refactor: update import paths to use Global.constants for consistency
