+++ src/shared/constants/Global.constants.ts
@@ -1,2 +1,5 @@
 export type ThemeMode = 'light' | 'dark';
 export const COOKIE_SESSION_KEY = 'accessToken'
+
+// Routes
+export const DASHBOARD_ROUTE = '/dashboard';
