+++ src/shared/types/Global.ts
@@ -1,6 +1,3 @@
 export interface YupError {
   message: string;
 }
-
-// Routes
-export const DASHBOARD_ROUTE = '/dashboard';
